% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excalidraw.R
\name{excalidraw-shiny}
\alias{excalidraw-shiny}
\alias{excalidrawOutput}
\alias{renderExcalidraw}
\alias{excalidraw_html}
\title{Shiny bindings for excalidraw}
\usage{
excalidrawOutput(outputId, width = "100\%", height = "400px")

renderExcalidraw(expr, env = parent.frame(), quoted = FALSE)

excalidraw_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a excalidraw}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using excalidraw within Shiny
applications and interactive Rmd documents.
}
